Added in scenes and manager
Content Update:

Added in Scenes-SceneManager.
Basically stripped down and refactored Game into an app running class so that we can progress back and forth, but moved the logic into Scenes and SceneManager.

Idea: Scene Manager will delegate downwards, focus on the scene currently active and update accordingly around it. Separates concern between scenes - Only minor flaw is I may need to refactor a bit later, will need to see how the procedural generation comes to fruition - may change between loads.

Added button components that call Lambdas to perform an action when clicked - only really used in Main Menu currently.

ToDo: Begin constructing the game world, And or the Settings - investigate modalization - overlay rather than a new screen or menu.

+++ Prototory SFML/Prototory SFML/Globals.h
@@ -1,10 +1,10 @@
 #pragma once
 #include <iostream>
+#include <SFML/Graphics.hpp>
 
 struct Globals
 {
-	static float constexpr SCREEN_WIDTH = 1000.f;
-	static float constexpr SCREEN_HEIGHT = 800.f;
-
+	static unsigned constexpr SCREEN_WIDTH = 1000.f;
+	static unsigned constexpr SCREEN_HEIGHT = 800.f;
 };
